requirements.txt
asgiref==3.7.2
Django==4.2.5
djangorestframework==3.14.0
joblib==1.3.2
numpy==1.25.2
pytz==2023.3.post1
scikit-learn==1.2.2
scipy==1.11.2
sqlparse==0.4.4
threadpoolctl==3.2.0
